+++ examples/resources/powerdns_network/import.sh
@@ -1,0 +1,1 @@
+terraform import powerdns_network.office 192.0.2.0/24