+++ examples/resources/powerdns_recursor_acl/import.sh
@@ -1,0 +1,1 @@
+terraform import powerdns_recursor_acl.clients allow-from