+++ examples/resources/powerdns_recursor_zone/import.sh
@@ -1,0 +1,1 @@
+terraform import powerdns_recursor_zone.internal corp.example.